Separate litmus tests and flit source code
The flit source code now compiles into a shared library instead
of an executable.  The test directory only needs a main.cpp that
consists of simply 4 lines of code:

#include flit.h
int main(int argCount, char* argList[]) {
  return runFlitTests(argCount, argList);
}
