--- src/MakefileDev
@@ -1,57 +1,0 @@
-#CC := clang++
-CC := g++
-FFLAGS ?=
-TARGET ?= runTests
-
-CPPFLAGS ?= -Wuninitialized -g #-fsanitize=thread
-CPPFLAGS += -IS3FP/
-CPPFLAGS += -DWFT=float -DSCH_LIB
-
-CPPFLAGS += $(FFLAGS)
-CPPFLAGS += -std=c++11
-CPPFLAGS += -Wno-shift-count-overflow
-CPPFLAGS += -Wall
-CPPFLAGS += -Wextra
-#CPPFLAGS += -Werror
-#CPPFLAGS += -MMD
-CPPFLAGS += -I.
-
-#for S3FP related test
-S3_REQUIRED 	:= -IS3FP/
-S3_REQUIRED 	+= -DWFT=float -DSCH_LIB
-
-CPPFLAGS += $(S3_REQUIRED)
-LINKFLAGS += -lm
-
-DEPFLAGS += -MD -MF $*.d
-
-TESTS    := $(wildcard tests/*.cpp)
-SOURCE   := $(wildcard *.cpp)
-SOURCE   += $(TESTS)
-
-OBJ      := $(SOURCE:.cpp=.o)
-DEPS     := $(SOURCE:.cpp=.d)
-
-
-.PHONY : all
-all :  $(TARGET)
-
-$(TARGET) : $(OBJ)
-	$(CC) $(CPPFLAGS) $(LINKFLAGS) -o $@ $^
-
-%.o : %.cpp
-	$(CC) $(CPPFLAGS) $(DEPFLAGS) -c $< -o $@
-
-.PRECIOUS: %.d
--include $(SOURCE:%.cpp=%.d)
-
-.PHONY : clean
-clean :
-	rm -f $(OBJ)
-	rm -f $(DEPS)
-
-.PHONY: veryclean distclean
-veryclean : distclean
-distclean : clean
-	rm -f $(TARGET)
-
